Now using module structure for transaction objects.
